Add files via upload
Good day sir. sorry for the late submission. I had some difficulty trying to configure the IDE to work with Git, but later on, I decided to manually upload the files. These are the class activities from the 8th of May, 2024 till date excluding the first committed activity; although, being part of the activities we have done within the same time frame.